+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Cache all calls to `os.getenv` via custom module [PR #231](https://github.com/3scale/apicast/pull/231)
 
 ### Fixed
+- [OAuth] Return correct state value back to client
+
 ### Removed
 
 ## [3.0.0-alpha1] - 2017-01-16